// packages/server/src/db/schema.ts
export const userRoles = ["waitlist", "user", "admin"] as const;
export type UserRole = (typeof userRoles)[number];

export const visibilityOptions = ["private", "team", "public"] as const;
export type VisibilityOption = (typeof visibilityOptions)[number];

export const user = sqliteTable("user", {
  id: text("id").primaryKey(),
  name: text("name").notNull(),
  username: text("username").notNull(),
  email: text("email").notNull().unique(),
  emailVerified: integer("email_verified", { mode: "boolean" }).default(false).notNull(),
  image: text("image"),
  role: text("role").$type<UserRole>().default("waitlist").notNull(),
  welcomeEmailSentAt: integer("welcome_email_sent_at", { mode: "timestamp_ms" }),
  createdAt: integer("created_at", { mode: "timestamp_ms" })
    .default(sql`(cast(unixepoch('subsecond') * 1000 as integer))`)
    .notNull(),
  updatedAt: integer("updated_at", { mode: "timestamp_ms" })
    .default(sql`(cast(unixepoch('subsecond') * 1000 as integer))`)
    .$onUpdate(() => /* @__PURE__ */ new Date())
    .notNull(),
});